% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc}
\alias{lc}
\title{lc}
\usage{
lc(o, p)
}
\arguments{
\item{o}{A numeric vector. Observed values.}

\item{p}{A numeric vector. Predicted values.}
}
\value{
Lack of correlation (LC)
}
\description{
Calculates the Lack of correlation (LC) from observed
and predicted values.
}
\details{
Interpretation: smaller is better.
}
\examples{
obs<-c(1:10)
pred<-c(1, 1 ,3, 2, 4, 5, 6, 8, 7, 10)
lc(o=obs, p=pred)

}
\references{
Gauch H. G., Hwang J. G., & Fick G. W. 2003. Model evaluation 
by comparison of model based predictions and measured values. Agronomy Journal, 
95(6), 1442-1446.

Piikki K., Wetterlind J., Soderstrom M., Stenberg B. (2021). Perspectives
on validation in digital soil mapping of continuous attributes. A review.
Soil Use and Management. \doi{10.1111/sum.12694}
}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom and Bo Stenberg, 
E-mail: \email{kristin.piikki@slu.se}
}
