% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mde.R
\name{mde}
\alias{mde}
\title{mde}
\usage{
mde(o, p)
}
\arguments{
\item{o}{A numeric vector. Observed values.}

\item{p}{A numeric vector. Predicted values.}
}
\value{
Median error (MdE)
}
\description{
Calculates the Median error (MdE) from observed and predicted 
values.
}
\details{
Interpretation: smaller is better. Similar to mean error (bias) but 
less sensitive to large errors. Sometimes called bias.
}
\examples{
obs<-c(1:10)
pred<-c(1, 1 ,3, 2, 4, 5, 6, 8, 7, 10)
mde(o=obs, p=pred)

}
\references{
Piikki K., Wetterlind J., Soderstrom M., Stenberg B. (2021). Perspectives
on validation in digital soil mapping of continuous attributes. A review.
Soil Use and Management. \doi{10.1111/sum.12694}
}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom and Bo Stenberg, 
E-mail: \email{kristin.piikki@slu.se}
}
