% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmse.R
\name{nmse}
\alias{nmse}
\title{nmse}
\usage{
nmse(o, p)
}
\arguments{
\item{o}{A numeric vector. Observed values.}

\item{p}{A numeric vector. Predicted values.}
}
\value{
Normalized mean squared error (NMSE)
}
\description{
Calculates the Normalized mean squared error (NMSE) from observed
and predicted values.
}
\details{
Interpretation: smaller is better.
}
\examples{
obs<-c(1:10)
pred<-c(1, 1 ,3, 2, 4, 5, 6, 8, 7, 10)
nmse(o=obs, p=pred)
}
\references{
Park S. J., Vlek P. L. G. 2002. Environmental correlation of 
three-dimensional soil spatial variability: a comparison of three adaptive 
techniques. Geoderma, 109(1-2), 117-140.
}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom and Bo Stenberg, 
E-mail: \email{kristin.piikki@slu.se}
}
