% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpd.R
\name{rpd}
\alias{rpd}
\title{rpd}
\usage{
rpd(o, p)
}
\arguments{
\item{o}{A numeric vector. Observed values.}

\item{p}{A numeric vector. Predicted values.}
}
\value{
Ratio of performance to deviation (RPD).
}
\description{
Calculates the Ratio of performance to deviation (RPD) from observed
and predicted values.
}
\details{
Interpretation: larger is better.
}
\examples{
obs<-c(1:10)
pred<-c(1, 1 ,3, 2, 4, 5, 6, 8, 7, 10)
rpd(o=obs, p=pred)

}
\references{
Piikki K., Wetterlind J., Soderstrom M., Stenberg B. (2021). Perspectives
on validation in digital soil mapping of continuous attributes. A review.
Soil Use and Management. \doi{10.1111/sum.12694}
}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom & Bo Stenberg
\email{kristin.piikki@slu.se}
}
