% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse}
\alias{sse}
\title{sse}
\usage{
sse(o, p)
}
\arguments{
\item{o}{A numeric vector. Observed values.}

\item{p}{A numeric vector. Predicted values.}
}
\value{
Sum of squares for error (SSE).
}
\description{
Calculates the sum of squares for error (SSE) from observed and
predicted values.
}
\details{
Interpretation: smaller is better.
}
\examples{
obs<-c(1:10)
pred<-c(1, 1 ,3, 2, 4, 5, 6, 8, 7, 10)
sse(o=obs, p=pred)

}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom and Bo Stenberg, 
E-mail: \email{kristin.piikki@slu.se}
}
