% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sst.R
\name{sst}
\alias{sst}
\title{sst}
\usage{
sst(o)
}
\arguments{
\item{o}{A numeric vector. Observed values.}
}
\value{
Total sums of squares (SST).
}
\description{
Calculates the Total sums of squares (SST) from a vector of
observed values.
}
\details{
Interpretation: smaller is better.
}
\examples{
obs<-c(1:10)
sst(o=obs)

}
\author{
Kristin Piikki, Johanna Wetterlind, Mats Soderstrom and Bo Stenberg, 
E-mail: \email{kristin.piikki@slu.se}
}
